Add missing headers. Implement a function to check that the parenthesis sequence is correct.

+++ data_structures/test_array.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <array>
 #include <vector>
+#include <string>
 #include "tools.h"
 
 template <size_t Size>
@@ -82,7 +83,7 @@ void array_rotation() {
 // Remove all zeros from array for O (n)
 // 1 0 2 0 0 3 0 -> 1 2 3
 void removeZeros(std::vector<int> data) {
-  int size = data.size();
+  size_t size = data.size();
   int zeroCount = 0;
   for (int i = 0; i < size; ++i) {
     if (data[i] == 0) {
